Add integration tests which query the actual implementation over HTTP.
This change adds a failing integration test for `/inspect`.
We spin up the axum router in the testing environment and bind to an ephemeral port.
The testing client then issues requests via `hyper` client.
Although this isn't testing `main.rs` itself, it's reasonably close.

To enable consistent integration tests, we have added a parameter to the agent to specify the root log directory.
The parameter is optional, with the expectation it will only be used during testing.
We update our assumptions to allow for the agent to be called with an overriden log root.
For example, a malicious actoring gaining access to the host could enable the agent with an unexpected path.