Implement reverse read for in-memory file-like struct.
This change focuses on the algorithm to process a file from the bottom upwards.
We implement the algorithm for a testable in-memory file-like struct, with the
intent of adding support for real files in a following change.
The basic algorithm is described in the readme.

Includes some key updates to the requirements and design:
* Log files are utf-8 encoded.
* Ultimately, we'll need to stream the results to minimize agent memory impact.