Asynchronously process log files and stream them over the request.
This change implements request streaming to minimize the memory overhead of the agent.
As log lines are made available, they are streamed out to the client immediately (still within a single HTTP body).

The trick to achieving this is to move the reverse read processing into a background task.
This background task puts log lines into a channel (a FIFO queue), and the request handler reads off the channel simulatenously.

Testing on my machine for the large ~2GB file shows a significant improvement
(~8 MB of memory and ~1.5 minutes vs the previous ~4 GB and ~2 minutes).